chore: remove unused file and redux store and move dir
