+++ src/components/NavigationBar/NavigationBar.tsx
@@ -3,7 +3,7 @@ import dynamic from 'next/dynamic';
 import Link from 'next/link';
 import { useRouter } from 'next/router';
 import styled from 'styled-components';
-import cssVar from '~/utils/cssVar';
+import cssVar from '~/lib/styles/cssVar';
 import Icon from '../icons';
 import { Item } from './Item';
