+++ src/lib/styles/cssVar.ts
@@ -1,5 +1,5 @@
 const cssVar = (variables: string) => {
-  return `var(--${variables})`;
+  return `var(--${variables.replace(/-/g, '_')})`;
 };
 
 export default cssVar;
