+++ src/lib/styles/theme.ts
@@ -1,0 +1,8 @@
+export const LightPalette = {};
+
+export const DarkPalette = {};
+
+export const Palette = {
+  light: LightPalette,
+  dark: DarkPalette,
+};
