+++ src/lib/styles/types.ts
@@ -1,0 +1,10 @@
+export type Theme = 'dark' | 'light';
+
+export interface Palette {
+  text1: string;
+  text2: string;
+  text3: string;
+  text4: string;
+  bg_page: string;
+  skeleton: string;
+}
