+++ src/store/store.ts
@@ -1,10 +1,8 @@
 import { configureStore } from '@reduxjs/toolkit';
 import modalReducer from '~/components/common/Modal/modalSlice';
-import themeReducer from '~styles/theme/themeSlice';
 
 const store = configureStore({
   reducer: {
-    isDarkMode: themeReducer,
     isModalOpen: modalReducer,
   },
 });
